@@ -13,28 +13,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.2",
-    "@angular/common": "^5.0.2",
-    "@angular/compiler": "^5.0.2",
-    "@angular/core": "^5.0.2",
-    "@angular/forms": "^5.0.2",
-    "@angular/http": "^5.0.2",
-    "@angular/platform-browser": "^5.0.2",
-    "@angular/platform-browser-dynamic": "^5.0.2",
-    "@angular/platform-server": "^5.0.2",
-    "@angular/router": "^5.0.2",
-    "angularfire2": "^5.0.0-rc.4",
+    "@angular-devkit/schematics": "^0.7.3",
+    "@angular/animations": "^6.1.2",
+    "@angular/common": "^6.1.2",
+    "@angular/core": "^6.1.2",
+    "@angular/forms": "^6.1.2",
+    "@angular/http": "^6.1.2",
+    "@angular/platform-browser": "^6.1.2",
+    "@angular/platform-browser-dynamic": "^6.1.2",
+    "@angular/platform-server": "^6.1.2",
+    "@angular/router": "^6.1.2",
+    "angularfire2": "^5.0.0-rc.11",
     "bulma": "^0.6.2",
     "core-js": "^2.5.1",
-    "firebase": "^4.6.2",
+    "firebase": "^5.3.1",
     "gsap": "^1.20.4",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.18"
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.2.2",
+    "webpack": "^4.16.5",
+    "webpack-dev-server": "^3.1.5",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.2",
-    "@angular/compiler-cli": "^5.0.2",
-    "@angular/language-service": "^5.0.2",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.3",
+    "@angular/compiler": "^6.1.2",
+    "@angular/compiler-cli": "^6.1.2",
+    "@angular/language-service": "^6.1.2",
     "@types/jasmine": "2.8.2",
     "@types/node": "~8.0.53",
     "codelyzer": "~4.0.1",
@@ -52,6 +57,6 @@
     "sw-precache-webpack-plugin": "^0.11.4",
     "ts-node": "~3.3.0",
     "tslint": "^5.8.0",
-    "typescript": "~2.4.2"
+    "typescript": "~2.9.2"
   }
 }
