@@ -3,11 +3,12 @@ import { Router } from '@angular/router';
 
 import * as firebase from 'firebase/app';
 import { AngularFireAuth } from 'angularfire2/auth';
+import { auth } from 'firebase';
 import { AngularFirestore, AngularFirestoreDocument } from 'angularfire2/firestore';
 import { NotifyService } from './shared/notifications/notify.service';
 
-import { Observable } from 'rxjs/Observable';
-import { switchMap } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
+import { switchMap, startWith, tap, filter } from 'rxjs/operators';
 
 interface User {
   uid: string;
@@ -21,19 +22,23 @@ export class AuthService {
 
   user: Observable<User | null>;
 
-  constructor(private afAuth: AngularFireAuth,
-              private afs: AngularFirestore,
-              private router: Router,
-              private notify: NotifyService) {
-
-    this.user = this.afAuth.authState
-      .switchMap((user) => {
+  constructor(
+    private afAuth: AngularFireAuth,
+    private afs: AngularFirestore,
+    private router: Router,
+    private notify: NotifyService
+  ) {
+    this.user = this.afAuth.authState.pipe(
+      switchMap(user => {
         if (user) {
           return this.afs.doc<User>(`users/${user.uid}`).valueChanges();
         } else {
-          return Observable.of(null);
+          return of(null);
         }
-      });
+      })
+      // tap(user => localStorage.setItem('user', JSON.stringify(user))),
+      // startWith(JSON.parse(localStorage.getItem('user')))
+    );
   }
 
   ////// OAuth Methods /////
@@ -67,17 +72,16 @@ export class AuthService {
       .catch((error) => this.handleError(error) );
   }
 
-  //// Anonymous Auth ////
 
+//// Anonymous Auth ////
   anonymousLogin() {
-    return this.afAuth.auth.signInAnonymously()
-      .then((user) => {
+    return this.afAuth.auth
+      .signInAnonymously()
+      .then(credential => {
         this.notify.update('Welcome to Firestarter!!!', 'success');
-        return this.updateUserData(user); // if using firestore
+        return this.updateUserData(credential.user); // if using firestore
       })
-      .catch((error) => {
-        console.error(error.code);
-        console.error(error.message);
+      .catch(error => {
         this.handleError(error);
       });
   }
@@ -85,35 +89,38 @@ export class AuthService {
   //// Email/Password Auth ////
 
   emailSignUp(email: string, password: string) {
-    return this.afAuth.auth.createUserWithEmailAndPassword(email, password)
-      .then((user) => {
-        this.notify.update('Welcome to Firestarter!!!', 'success');
-        return this.updateUserData(user); // if using firestore
+    return this.afAuth.auth
+      .createUserWithEmailAndPassword(email, password)
+      .then(credential => {
+        this.notify.update('Welcome new user!', 'success');
+        return this.updateUserData(credential.user); // if using firestore
       })
-      .catch((error) => this.handleError(error) );
+      .catch(error => this.handleError(error));
   }
 
   emailLogin(email: string, password: string) {
-    return this.afAuth.auth.signInWithEmailAndPassword(email, password)
-      .then((user) => {
-        this.notify.update('Welcome to Firestarter!!!', 'success')
-        return this.updateUserData(user); // if using firestore
+    return this.afAuth.auth
+      .signInWithEmailAndPassword(email, password)
+      .then(credential => {
+        this.notify.update('Welcome back!', 'success');
+        return this.updateUserData(credential.user);
       })
-      .catch((error) => this.handleError(error) );
+      .catch(error => this.handleError(error));
   }
 
   // Sends email allowing user to reset password
   resetPassword(email: string) {
-    const fbAuth = firebase.auth();
+    const fbAuth = auth();
 
-    return fbAuth.sendPasswordResetEmail(email)
+    return fbAuth
+      .sendPasswordResetEmail(email)
       .then(() => this.notify.update('Password update email sent', 'info'))
-      .catch((error) => this.handleError(error));
+      .catch(error => this.handleError(error));
   }
 
   signOut() {
     this.afAuth.auth.signOut().then(() => {
-        this.router.navigate(['/']);
+      this.router.navigate(['/']);
     });
   }
 
